Got SLAM to compile
Built hpp file for slam_map_t

Made mapper publishing actually work

+++ src/GridMap.hpp
@@ -41,14 +41,14 @@ public:
 		return (x_idx + y_idx * cells_per_row);
 	}
 
-	uint8_t at(double x, double y) const
+	int16_t at(double x, double y) const
 	{
 		std::size_t idx = convertToGridCoords(x,y);
 		return map[idx];
 	}
 
 	//accesses the vector 
-	uint8_t& at(double x, double y)
+	int16_t& at(double x, double y)
 	{
 		std::size_t idx = convertToGridCoords(x,y);
 		return map[idx];
@@ -83,12 +83,12 @@ public:
 		return map.size();
 	}
 
-	const uint8_t& operator[](std::size_t idx) const 
+	const int16_t& operator[](std::size_t idx) const 
 	{
 		return map[idx];
 	}
 
-	uint8_t& operator[](std::size_t idx)
+	int16_t& operator[](std::size_t idx)
 	{
 		return map[idx];
 	}
@@ -99,7 +99,7 @@ public:
 private:
 	//actual map containing the probabilities 
 	//idx 0 is bottom left (-y, -x)
-	std::vector<uint8_t> map;
+	std::vector<int16_t> map;
 
 	//extents of the map in x and y
 	double max_x;
